Added single quotes around the executable path (#1017)
First noticed in:
https://github.com/rust-lang/rust-analyzer/issues/11489

where I saw that generics prevent a test case from running. It turns out
that the `<` and `>` made my shell I was redirecting to a file (I tested
this in other shells and got the same behavior). As such, I made a PR to
rust-analyzer, but it turns out that this is an "IDE" bug, as we
directly send the cmd to be ran on the terminal.

To fix the issue I simply surrounded the path with quotes.