Use the correct executable for debugging (#1055)
The previous version of the code just uses the first executable
it finds in the JSON output of cargo. This is problematic because
multiple executables might be produced, for example when debugging
an integration test, and the order of these artifacts in the JSON
output can change from one invocation of cargo to the next.

Analyze the arguments given to cargo instead, from which the
expected values for target.kind and target.name can be extracted.
With this information, the executable should be uniquely identifiable.